secure type of function return

+++ electroacoustical.py
@@ -195,7 +195,7 @@ def wind_coil(wire: Wire,
         "Calculate the number of windings that fit on one layer of coil."
         # 1 winding less on each stacked layer if stacking coefficient is less than or equal to 0.9
         n_winding = h_winding_target / wire.h_avg - i_layer * reduce_per_layer
-        return round(n_winding)
+        return int(round(n_winding))
 
     N_windings = [N_winding_for_single_layer(i_layer) for i_layer in range(N_layers)]
     if any([N_winding < 1 for N_winding in N_windings]):
